# Remove pointer indirection on chunk bstream