It isn't necessary since the member is never changed after initialization.

Saves 8 bytes per chunk, and a bit of time following the pointer and garbage-collecting the extra object.