@krasi-georgiev 2019-01-18
it seems that this has some positive gains as well

```
benchmark                                                                                                              old ns/op      new ns/op      delta
BenchmarkRangeQuery/expr=rate(a_hundred[1d]),steps=1000-48                                                             1393708393     1105947109     -20.65%
BenchmarkRangeQuery/expr=rate(a_ten[1d]),steps=1000-48                                                                 138105029      110480466      -20.00%
BenchmarkRangeQuery/expr=rate(a_hundred[1d]),steps=100-48                                                              224942358      190443003      -15.34%
BenchmarkRangeQuery/expr=rate(a_one[1d]),steps=1000-48                                                                 14051544       11958272       -14.90%
BenchmarkRangeQuery/expr=rate(a_ten[1d]),steps=100-48                                                                  22394837       19175200       -14.38%
BenchmarkRangeQuery/expr=rate(a_one[1d]),steps=100-48                                                                  2444610        2144917        -12.26%
BenchmarkRangeQuery/expr=changes(a_hundred[1d]),steps=10-48                                                            117204916      108713575      -7.24%
BenchmarkRangeQuery/expr=rate(a_hundred[1d]),steps=10-48                                                               106870749      99335234       -7.05%
BenchmarkRangeQuery/expr=rate(a_ten[1d]),steps=10-48                                                                   10685653       9976464        -6.64%
BenchmarkRangeQuery/expr=rate(a_one[1d]),steps=10-48                                                                   1283516        1202529        -6.31%
BenchmarkRangeQuery/expr=rate(a_one[1d]),steps=1-48                                                                    1121137        1070152        -4.55%
BenchmarkRangeQuery/expr=changes(a_one[1d]),steps=1-48                                                                 1136138        1092632        -3.83%
BenchmarkRangeQuery/expr=changes(a_ten[1d]),steps=1-48                                                                 9461996        9116670        -3.65%
BenchmarkRangeQuery/expr=rate(a_hundred[1d]),steps=1-48                                                                92733291       89427747       -3.56%
BenchmarkRangeQuery/expr=changes(a_one[1d]),steps=10-48                                                                1352822        1311888        -3.03%
BenchmarkRangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=1-48                                                  1388812        1346957        -3.01%
BenchmarkRangeQuery/expr=a_ten,steps=1000-48                                                                           1340605        1300413        -3.00%
BenchmarkRangeQuery/expr=sum(a_ten),steps=1-48                                                                         138604         134872         -2.69%
BenchmarkRangeQuery/expr=rate(a_one[1m])_+_rate(b_one[1m]),steps=1000-48                                               1352684        1316923        -2.64%
BenchmarkRangeQuery/expr=sum_without_(l)(rate(a_one[1m])),steps=100-48                                                 185766         181162         -2.48%
BenchmarkRangeQuery/expr=changes(a_hundred[1d]),steps=1-48                                                             88933163       91188969       +2.54%
BenchmarkRangeQuery/expr=changes(a_ten[1d]),steps=10-48                                                                11246240       10973807       -2.42%
BenchmarkRangeQuery/expr=sum_without_(l)(rate(a_hundred[1m]))_/_sum_without_(l)(rate(b_hundred[1m])),steps=1000-48     64364632       62825183       -2.39%
BenchmarkRangeQuery/expr=rate(a_ten[1d]),steps=1-48                                                                    9228732        9008127        -2.39%
BenchmarkRangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=1-48                                              102687379      100278388      -2.35%
BenchmarkRangeQuery/expr=sum_without_(l)(rate(a_ten[1m]))_/_sum_without_(l)(rate(b_ten[1m])),steps=1000-48             9010018        8804233        -2.28%
BenchmarkRangeQuery/expr=a_ten_unless_b_ten{l=~'.*[0-4]$'},steps=10-48                                                 550437         538196         -2.22%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=10000-48                                                                  11966178       11706762       -2.17%
BenchmarkRangeQuery/expr=a_one_unless_b_one{l=~'.*[0-4]$'},steps=10-48                                                 352875         345316         -2.14%
BenchmarkRangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=10-48                                             163522962      160099379      -2.09%
BenchmarkRangeQuery/expr=label_replace(a_one,_'l2',_'$1',_'l',_'(.*)'),steps=10-48                                     77211          78789          +2.04%
BenchmarkRangeQuery/expr=-a_one,steps=1000-48                                                                          161392         158283         -1.93%
BenchmarkRangeQuery/expr=sum(a_one),steps=1-48                                                                         43614          42782          -1.91%
BenchmarkRangeQuery/expr=a_ten_or_b_ten{l=~'.*[0-4]$'},steps=1-48                                                      510913         501199         -1.90%
BenchmarkRangeQuery/expr=sum_without_(le)(h_one),steps=1000-48                                                         3612202        3544084        -1.89%
BenchmarkRangeQuery/expr=label_join(a_ten,_'l2',_'-',_'l',_'l'),steps=1000-48                                          6772759        6654517        -1.75%
BenchmarkRangeQuery/expr=rate(a_one[1m]),steps=100-48                                                                  73553          72296          -1.71%
BenchmarkRangeQuery/expr=-a_one,steps=1-48                                                                             39745          39067          -1.71%
BenchmarkRangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=1-48                                                     505627         497117         -1.68%
BenchmarkRangeQuery/expr=a_one,steps=1000-48                                                                           154029         156664         +1.71%
BenchmarkRangeQuery/expr=sum(a_hundred),steps=1-48                                                                     1060371        1042859        -1.65%
BenchmarkRangeQuery/expr=abs(a_one),steps=100-48                                                                       95351          96937          +1.66%
BenchmarkRangeQuery/expr=sum_without_(l)(rate(a_one[1m]))_/_sum_without_(l)(rate(b_one[1m])),steps=100-48              434193         427204         -1.61%
BenchmarkRangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=100-48                                                   400599         394166         -1.61%
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=10000-48                                                                25240479       24837133       -1.60%
BenchmarkRangeQuery/expr=changes(a_hundred[1d]),steps=100-48                                                           285373016      280861783      -1.58%
BenchmarkRangeQuery/expr=sum_without_(l)(rate(a_one[1m]))_/_sum_without_(l)(rate(b_one[1m])),steps=1000-48             3405163        3351462        -1.58%
BenchmarkRangeQuery/expr=sum_without_(l)(rate(a_hundred[1m])),steps=100-48                                             4627291        4554343        -1.58%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=1-48                                                                      62343          61373          -1.56%
BenchmarkRangeQuery/expr=label_replace(a_ten,_'l2',_'$1',_'l',_'(.*)'),steps=1-48                                      189102         186200         -1.53%
BenchmarkRangeQuery/expr=rate(a_one[1m])_+_rate(b_one[1m]),steps=100-48                                                210918         207719         -1.52%
BenchmarkRangeQuery/expr=a_one_or_b_one{l=~'.*[0-4]$'},steps=1000-48                                                   1086024        1069751        -1.50%
BenchmarkRangeQuery/expr=a_hundred_unless_b_hundred{l=~'.*[0-4]$'},steps=1-48                                          2012809        1983140        -1.47%
BenchmarkRangeQuery/expr=sum_without_(le)(h_hundred),steps=100-48                                                      55242853       54431545       -1.47%
BenchmarkRangeQuery/expr=sum_without_(l)(rate(a_ten[1m])),steps=1000-48                                                4368246        4304481        -1.46%
BenchmarkRangeQuery/expr=label_join(a_one,_'l2',_'-',_'l',_'l'),steps=1000-48                                          1029889        1014894        -1.46%
BenchmarkRangeQuery/expr=sum_by_(l)(h_hundred),steps=1000-48                                                           392205522      397966568      +1.47%
BenchmarkRangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=10-48                                                    347546         342552         -1.44%
BenchmarkRangeQuery/expr=changes(a_one[1d]),steps=100-48                                                               3076067        3032107        -1.43%
BenchmarkRangeQuery/expr=sum_without_(l)(h_one),steps=1000-48                                                          14271870       14069401       -1.42%
BenchmarkRangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=1-48                                                     346851         341990         -1.40%
BenchmarkRangeQuery/expr=-a_ten,steps=1000-48                                                                          1358326        1339356        -1.40%
BenchmarkRangeQuery/expr=abs(a_one),steps=1-48                                                                         44442          45065          +1.40%
BenchmarkRangeQuery/expr=sum_without_(l)(rate(a_ten[1m]))_/_sum_without_(l)(rate(b_ten[1m])),steps=100-48              1204517        1188022        -1.37%
BenchmarkRangeQuery/expr=a_ten_-_b_ten,steps=1-48                                                                      272229         268504         -1.37%
BenchmarkRangeQuery/expr=label_join(a_hundred,_'l2',_'-',_'l',_'l'),steps=1-48                                         1309636        1292150        -1.34%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=10-48                                                                     74610          73627          -1.32%
BenchmarkRangeQuery/expr=sum_without_(l)(rate(a_hundred[1m]))_/_sum_without_(l)(rate(b_hundred[1m])),steps=1-48        2323816        2293402        -1.31%
BenchmarkRangeQuery/expr=sum_by_(l)(h_hundred),steps=10-48                                                             15231568       15034206       -1.30%
BenchmarkRangeQuery/expr=sum_without_(l)(h_ten),steps=1-48                                                             1185207        1170187        -1.27%
BenchmarkRangeQuery/expr=sum_without_(l)(rate(a_hundred[1m]))_/_sum_without_(l)(rate(b_hundred[1m])),steps=10-48       2770367        2735692        -1.25%
BenchmarkRangeQuery/expr=sum_without_(l)(h_one),steps=10-48                                                            298562         294830         -1.25%
BenchmarkRangeQuery/expr=sum_without_(l)(h_hundred),steps=1-48                                                         12613463       12456064       -1.25%
BenchmarkRangeQuery/expr=-a_ten,steps=1-48                                                                             134300         132625         -1.25%
BenchmarkRangeQuery/expr=abs(a_one),steps=1000-48                                                                      531986         538623         +1.25%
BenchmarkRangeQuery/expr=a_one,steps=100-48                                                                            47634          48228          +1.25%
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=1000-48                                                                 2336201        2307603        -1.22%
BenchmarkRangeQuery/expr=abs(a_hundred),steps=1000-48                                                                  59271699       59996139       +1.22%
BenchmarkRangeQuery/expr=histogram_quantile(0.9,_rate(h_one[5m])),steps=100-48                                         823170         813325         -1.20%
BenchmarkRangeQuery/expr=sum(a_hundred),steps=1000-48                                                                  21957367       22220510       +1.20%
BenchmarkRangeQuery/expr=sum_without_(le)(h_ten),steps=10-48                                                           1460066        1442842        -1.18%
BenchmarkRangeQuery/expr=sum_without_(l)(rate(a_ten[1m]))_/_sum_without_(l)(rate(b_ten[1m])),steps=1-48                301384         297880         -1.16%
BenchmarkRangeQuery/expr=sum_without_(l)(rate(a_hundred[1m])),steps=10-48                                              1418750        1402326        -1.16%
BenchmarkRangeQuery/expr=a_ten_-_b_ten,steps=10000-48                                                                  126497833      125038899      -1.15%
BenchmarkRangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=100-48                                                   1100062        1087514        -1.14%
BenchmarkRangeQuery/expr=rate(a_one[1m])_+_rate(b_one[1m]),steps=1-48                                                  80696          79782          -1.13%
BenchmarkRangeQuery/expr=sum_without_(l)(rate(a_hundred[1m]))_/_sum_without_(l)(rate(b_hundred[1m])),steps=100-48      8709789        8611189        -1.13%
BenchmarkRangeQuery/expr=-a_ten,steps=10-48                                                                            138410         136849         -1.13%
BenchmarkRangeQuery/expr=sum_without_(l)(h_ten),steps=10-48                                                            1491573        1475272        -1.09%
BenchmarkRangeQuery/expr=sum_without_(l)(rate(a_hundred[1m])),steps=1000-48                                            34625486       34248547       -1.09%
BenchmarkRangeQuery/expr=rate(a_one[1m]),steps=10-48                                                                   48396          47876          -1.07%
BenchmarkRangeQuery/expr=abs(a_ten),steps=10-48                                                                        198385         200520         +1.08%
BenchmarkRangeQuery/expr=abs(a_hundred),steps=100-48                                                                   6922866        6994474        +1.03%
BenchmarkRangeQuery/expr=sum_by_(l)(h_one),steps=1-48                                                                  149933         148403         -1.02%
BenchmarkRangeQuery/expr=label_replace(a_ten,_'l2',_'$1',_'l',_'(.*)'),steps=100-48                                    814037         822292         +1.01%
BenchmarkRangeQuery/expr=label_join(a_ten,_'l2',_'-',_'l',_'l'),steps=10-48                                            225516         223297         -0.98%
BenchmarkRangeQuery/expr=sum_by_(le)(h_one),steps=1-48                                                                 170486         168818         -0.98%
BenchmarkRangeQuery/expr=rate(a_ten[1m])_+_rate(b_ten[1m]),steps=1000-48                                               13792640       13657750       -0.98%
BenchmarkRangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=1-48                                             2006371        1986829        -0.97%
BenchmarkRangeQuery/expr=sum_without_(le)(h_hundred),steps=10-48                                                       15779813       15626799       -0.97%
BenchmarkRangeQuery/expr=histogram_quantile(0.9,_rate(h_one[5m])),steps=1-48                                           188875         187055         -0.96%
BenchmarkRangeQuery/expr=sum_without_(l)(h_hundred),steps=1000-48                                                      358986603      362472891      +0.97%
BenchmarkRangeQuery/expr=rate(a_one[1m])_+_rate(b_one[1m]),steps=10-48                                                 90731          89859          -0.96%
BenchmarkRangeQuery/expr=sum_by_(le)(h_ten),steps=1-48                                                                 1168708        1157518        -0.96%
BenchmarkRangeQuery/expr=a_hundred_or_b_hundred{l=~'.*[0-4]$'},steps=10-48                                             2724566        2698594        -0.95%
BenchmarkRangeQuery/expr=sum_without_(le)(h_ten),steps=1-48                                                            1177575        1166389        -0.95%
BenchmarkRangeQuery/expr=histogram_quantile(0.9,_rate(h_ten[5m])),steps=1-48                                           1413120        1399732        -0.95%
BenchmarkRangeQuery/expr=sum_by_(l)(h_ten),steps=1-48                                                                  1164533        1153511        -0.95%
BenchmarkRangeQuery/expr=abs(a_hundred),steps=1-48                                                                     1243401        1231854        -0.93%
BenchmarkRangeQuery/expr=histogram_quantile(0.9,_rate(h_hundred[5m])),steps=1000-48                                    686702572      680330652      -0.93%
BenchmarkRangeQuery/expr=sum_without_(le)(h_hundred),steps=1000-48                                                     452426128      448272019      -0.92%
BenchmarkRangeQuery/expr=a_hundred_or_b_hundred{l=~'.*[0-4]$'},steps=100-48                                            10262183       10355097       +0.91%
BenchmarkRangeQuery/expr=sum_by_(l)(h_one),steps=10-48                                                                 168371         166863         -0.90%
BenchmarkRangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=10-48                                                    547963         543057         -0.90%
BenchmarkRangeQuery/expr=a_one_or_b_one{l=~'.*[0-4]$'},steps=1-48                                                      343053         339986         -0.89%
BenchmarkRangeQuery/expr=label_replace(a_ten,_'l2',_'$1',_'l',_'(.*)'),steps=10-48                                     236323         238401         +0.88%
BenchmarkRangeQuery/expr=histogram_quantile(0.9,_rate(h_ten[5m])),steps=100-48                                         7755938        7689170        -0.86%
BenchmarkRangeQuery/expr=sum_without_(le)(h_one),steps=100-48                                                          503998         499677         -0.86%
BenchmarkRangeQuery/expr=sum_by_(l)(h_ten),steps=10-48                                                                 1393556        1381662        -0.85%
BenchmarkRangeQuery/expr=a_one,steps=1-48                                                                              36516          36215          -0.82%
BenchmarkRangeQuery/expr=label_join(a_one,_'l2',_'-',_'l',_'l'),steps=1-48                                             57739          58218          +0.83%
BenchmarkRangeQuery/expr=-a_hundred,steps=1-48                                                                         1093044        1084058        -0.82%
BenchmarkRangeQuery/expr=sum_without_(l)(h_hundred),steps=10-48                                                        14710880       14592630       -0.80%
BenchmarkRangeQuery/expr=histogram_quantile(0.9,_rate(h_one[5m])),steps=10-48                                          239679         237780         -0.79%
BenchmarkRangeQuery/expr=a_one_unless_b_one{l=~'.*[0-4]$'},steps=100-48                                                411965         415244         +0.80%
…